Added currency attribute to related entities, refactoring

+++ src/main/java/ru/tink/practice/service/external/exchange/ExternalExchangeService.java
@@ -2,6 +2,7 @@ package ru.tink.practice.service.external.exchange;
 
 import ru.tink.practice.dto.external.moex.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -12,7 +13,7 @@ public interface ExternalExchangeService {
 
     SecurityFullInfoDTO getSecurityBySecid(String secid);
 
-    Double getCurrentSecurityPrice(String secid, String securityType);
+    BigDecimal getCurrentSecurityPrice(String secid, String securityType);
 
     List<CurrentPriceDTO> getPriceStatisticsByNumberOfDays(String secid, Map<String, String> params);
